backend/requirements.txt
fastapi
uvicorn 
openai
python-multipart
Pillow
pdf2image
Pillow-HEIF
google-generativeai
python-dotenv
python-docx